Fixes NDRUNION unpacking inside a pointer to an array of union elements
An example of such scenario is [MS-DHCPE], `R_DhcpEnumSubnetElementsV4` call
(section 3.1.4.31).
The returned element is a `LPDHCP_SUBNET_ELEMENT_INFO_ARRAY_V4`:
```
typedef struct _DHCP_SUBNET_ELEMENT_INFO_ARRAY_V4 {
  DWORD NumElements;
[size_is(NumElements)] LPDHCP_SUBNET_ELEMENT_DATA_V4 Elements;
} DHCP_SUBNET_ELEMENT_INFO_ARRAY_V4,
*LPDHCP_SUBNET_ELEMENT_INFO_ARRAY_V4;
```
each element of that array is a union like this:
```
&& (DhcpIpRangesDhcpOnly <= (E)))?(0):(E))
typedef struct _DHCP_SUBNET_ELEMENT_DATA_V4 {
  DHCP_SUBNET_ELEMENT_TYPE ElementType;
  [switch_is(ELEMENT_MASK(ElementType)), switch_type(DHCP_SUBNET_ELEMENT_TYPE)]
  union _DHCP_SUBNET_ELEMENT_UNION_V4 {
    [case(DhcpIpRanges)]
    DHCP_IP_RANGE* IpRange; [case(DhcpSecondaryHosts)]
    DHCP_HOST_INFO* SecondaryHost; [case(DhcpReservedIps)]
    DHCP_IP_RESERVATION_V4* ReservedIp; [case(DhcpExcludedIpRanges)]
    DHCP_IP_RANGE* ExcludeIpRange; [case(DhcpIpUsedClusters)]
    DHCP_IP_CLUSTER* IpUsedCluster;
  } Element;
} DHCP_SUBNET_ELEMENT_DATA_V4, *LPDHCP_SUBNET_ELEMENT_DATA_V4;
```

Trying to decode a structure like this would trigger an error.

Main problem was the NDRUNION constructor was interpreting all the
pointers inside the structure as NDR top level pointers, which is
not always true (only true if UNION belongs to the NDRCALL directly).

Changing this behaviour seems to fix this issue while mantaining
the rest of the test cases working. More test is needed tho.