Fixes https://github.com/CoreSecurity/impacket/issues/66 reported by deniskhodus.
We're now testing the `securityBindings` buffer is at least a two bytes long.
This shouldn't happen in theory, but maybe there are a few remaining bytes
left (probably padding bytes) in the last round of that loop.

Maybe we should also check if there could be 3 or more bytes left on that buffer,
but not long enough to fulfill a `SECURITYBINDING` structure.
If so, then there will be an exception trying to build a `SECURITYBINDING`
struct. However that error hasn't been reported.