Added extra check when parsing FIND_FIRST2 responses
- Samba 3.6.9 doesn't set record['NextEntryOffset'] to 0 for the last
record. Don't know exactly why, 'cause [MS-CIFS] clearly states that:

`NextEntryOffset: (4 bytes): This field contains the offset, in bytes, from this entry in the list to the next entry in the list. If there are no additional entries, the value MUST be zero (0x00000000).`

Now we're checking we reached the end of the buffer, just in case.